requirements.txt
logbook
requests
python-redmine
pypandoc
BeautifulSoup4
